requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
evaluate>=0.4.0
rouge_score>=0.1.2
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.14.0
tensorboard>=2.14.0
umap-learn>=0.5.3
hdbscan>=0.8.29
accelerate>=0.26.0 